requirements.txt
opencv-python
numpy
pillow
flask
flask-cors
firebase-admin
mtcnn
deepface
scipy
mediapipe
tensorflow